Start an initial `default_compiler_draft3.h` container
