+++ src/compiler/default_compiler_draft3.h
@@ -1,0 +1,1727 @@
+#ifndef SOURCEMETA_BLAZE_COMPILER_DEFAULT_COMPILER_DRAFT3_H_
+#define SOURCEMETA_BLAZE_COMPILER_DEFAULT_COMPILER_DRAFT3_H_
+
+#include <sourcemeta/blaze/compiler.h>
+#include <sourcemeta/blaze/evaluator.h>
+
+#include <sourcemeta/core/regex.h>
+
+#include <algorithm> // std::sort, std::ranges::any_of, std::ranges::all_of, std::find_if, std::ranges::none_of
+#include <cassert>   // assert
+#include <set>       // std::set
+#include <utility>   // std::move, std::to_underlying
+
+#include "compile_helpers.h"
+
+static auto parse_regex(const std::string &pattern,
+                        const sourcemeta::core::URI &base,
+                        const sourcemeta::core::WeakPointer &schema_location)
+    -> sourcemeta::core::Regex {
+  const auto result{sourcemeta::core::to_regex(pattern)};
+  if (!result.has_value()) [[unlikely]] {
+    throw sourcemeta::blaze::CompilerInvalidRegexError(
+        base, to_pointer(schema_location), pattern);
+  }
+
+  return result.value();
+}
+
+static auto
+relative_schema_location_size(const sourcemeta::blaze::Context &context,
+                              const sourcemeta::blaze::Instruction &step)
+    -> std::size_t {
+  return context.extra[step.extra_index].relative_schema_location.size();
+}
+
+static auto
+defines_direct_enumeration(const sourcemeta::blaze::Instructions &steps)
+    -> std::optional<std::size_t> {
+  const auto iterator{std::ranges::find_if(steps, [](const auto &step) {
+    return step.type == sourcemeta::blaze::InstructionIndex::AssertionEqual ||
+           step.type == sourcemeta::blaze::InstructionIndex::AssertionEqualsAny;
+  })};
+
+  if (iterator == steps.cend()) {
+    return std::nullopt;
+  }
+
+  return std::distance(steps.cbegin(), iterator);
+}
+
+static auto is_inside_disjunctor(const sourcemeta::core::WeakPointer &pointer)
+    -> bool {
+  return pointer.size() > 2 && pointer.at(pointer.size() - 2).is_index() &&
+         pointer.at(pointer.size() - 3).is_property() &&
+         (pointer.at(pointer.size() - 3).to_property() == "oneOf" ||
+          pointer.at(pointer.size() - 3).to_property() == "anyOf");
+}
+
+static auto json_array_to_string_set(const sourcemeta::core::JSON &document)
+    -> sourcemeta::blaze::ValueStringSet {
+  sourcemeta::blaze::ValueStringSet result;
+  for (const auto &value : document.as_array()) {
+    assert(value.is_string());
+    result.insert(value.to_string());
+  }
+
+  return result;
+}
+
+static auto
+is_closed_properties_required(const sourcemeta::core::JSON &schema,
+                              const sourcemeta::blaze::ValueStringSet &required)
+    -> bool {
+  return !schema.defines("patternProperties") &&
+         schema.defines("additionalProperties") &&
+         schema.at("additionalProperties").is_boolean() &&
+         !schema.at("additionalProperties").to_boolean() &&
+         schema.defines("properties") && schema.at("properties").is_object() &&
+         schema.at("properties").size() == required.size() &&
+         std::ranges::all_of(required, [&schema](const auto &property) {
+           return schema.at("properties")
+               .defines(property.first, property.second);
+         });
+}
+
+static auto
+compile_properties(const sourcemeta::blaze::Context &context,
+                   const sourcemeta::blaze::SchemaContext &schema_context,
+                   const sourcemeta::blaze::DynamicContext &dynamic_context,
+                   const sourcemeta::blaze::Instructions &)
+    -> std::vector<std::pair<std::string, sourcemeta::blaze::Instructions>> {
+  std::vector<std::pair<std::string, sourcemeta::blaze::Instructions>>
+      properties;
+  for (const auto &entry : schema_context.schema.at("properties").as_object()) {
+    properties.emplace_back(
+        entry.first,
+        compile(context, schema_context, dynamic_context,
+                sourcemeta::blaze::make_weak_pointer(entry.first),
+                sourcemeta::blaze::make_weak_pointer(entry.first)));
+  }
+
+  // In many cases, `properties` have some subschemas that are small
+  // and some subschemas that are large. To attempt to improve performance,
+  // we prefer to evaluate smaller subschemas first, in the hope of failing
+  // earlier without spending a lot of time on other subschemas
+  if (context.tweaks.properties_reorder) {
+    std::ranges::sort(properties, [&context](const auto &left,
+                                             const auto &right) {
+      const auto left_size{recursive_template_size(left.second)};
+      const auto right_size{recursive_template_size(right.second)};
+      if (left_size == right_size) {
+        const auto left_direct_enumeration{
+            defines_direct_enumeration(left.second)};
+        const auto right_direct_enumeration{
+            defines_direct_enumeration(right.second)};
+
+        // Enumerations always take precedence
+        if (left_direct_enumeration.has_value() &&
+            right_direct_enumeration.has_value()) {
+          // If both options have a direct enumeration, we choose
+          // the one with the shorter relative schema location
+          return relative_schema_location_size(
+                     context, left.second.at(left_direct_enumeration.value())) <
+                 relative_schema_location_size(
+                     context,
+                     right.second.at(right_direct_enumeration.value()));
+        } else if (left_direct_enumeration.has_value()) {
+          return true;
+        } else if (right_direct_enumeration.has_value()) {
+          return false;
+        }
+
+        return left.first < right.first;
+      } else {
+        return left_size < right_size;
+      }
+    });
+  }
+
+  return properties;
+}
+
+static auto to_string_hashes(
+    std::vector<std::pair<sourcemeta::blaze::ValueString,
+                          sourcemeta::blaze::ValueStringSet::hash_type>>
+        &hashes) -> sourcemeta::blaze::ValueStringHashes {
+  assert(!hashes.empty());
+  std::ranges::sort(hashes, [](const auto &left, const auto &right) {
+    return left.first.size() < right.first.size();
+  });
+
+  sourcemeta::blaze::ValueStringHashes result;
+  // The idea with the table of contents is as follows: each index
+  // marks the starting and end positions for a string where the size
+  // is equal to the index.
+  result.second.resize(hashes.back().first.size() + 1, std::make_pair(0, 0));
+  // TODO(C++23): Use std::views::enumerate when available in libc++
+  for (std::size_t index = 0; index < hashes.size(); index++) {
+    result.first.emplace_back(hashes[index].second, hashes[index].first);
+    const auto string_size{hashes[index].first.size()};
+    // We leave index 0 to represent the empty string
+    const auto position{index + 1};
+    const auto lower_bound{
+        result.second[string_size].first == 0
+            ? position
+            : std::min(result.second[string_size].first, position)};
+    const auto upper_bound{
+        result.second[string_size].second == 0
+            ? position
+            : std::max(result.second[string_size].second, position)};
+    assert(lower_bound <= upper_bound);
+    assert(lower_bound > 0 && upper_bound > 0);
+    assert(string_size < result.second.size());
+    result.second[string_size] = std::make_pair(lower_bound, upper_bound);
+  }
+
+  assert(result.second.size() == hashes.back().first.size() + 1);
+  return result;
+}
+
+namespace internal {
+using namespace sourcemeta::blaze;
+
+auto compiler_draft3_core_ref(const Context &context,
+                              const SchemaContext &schema_context,
+                              const DynamicContext &dynamic_context,
+                              const Instructions &) -> Instructions {
+  const auto &entry{static_frame_entry(context, schema_context)};
+  const auto type{sourcemeta::core::SchemaReferenceType::Static};
+  const auto reference{context.frame.reference(type, entry.pointer)};
+  if (!reference.has_value()) [[unlikely]] {
+    throw sourcemeta::core::SchemaReferenceError(
+        schema_context.schema.at(dynamic_context.keyword).to_string(),
+        to_pointer(schema_context.relative_pointer),
+        "Could not resolve schema reference");
+  }
+
+  const auto key{std::make_tuple(type,
+                                 std::string_view{reference->get().destination},
+                                 schema_context.is_property_name)};
+  assert(context.targets.contains(key));
+  return {make(sourcemeta::blaze::InstructionIndex::ControlJump, context,
+               schema_context, dynamic_context,
+               ValueUnsignedInteger{context.targets.at(key).first})};
+}
+
+// There are two ways to compile `properties` depending on whether
+// most of the properties are marked as required using `required`
+// or whether most of the properties are optional. Each shines
+// in the corresponding case.
+auto properties_as_loop(const Context &context,
+                        const SchemaContext &schema_context,
+                        const sourcemeta::core::JSON &properties) -> bool {
+  if (context.tweaks.properties_always_unroll) {
+    return false;
+  }
+
+  using Known = sourcemeta::core::Vocabularies::Known;
+  const auto size{properties.size()};
+  const auto imports_validation_vocabulary =
+      schema_context.vocabularies.contains(Known::JSON_Schema_Draft_4) ||
+      schema_context.vocabularies.contains(Known::JSON_Schema_Draft_6) ||
+      schema_context.vocabularies.contains(Known::JSON_Schema_Draft_7) ||
+      schema_context.vocabularies.contains(
+          Known::JSON_Schema_2019_09_Validation) ||
+      schema_context.vocabularies.contains(
+          Known::JSON_Schema_2020_12_Validation);
+  const auto imports_const =
+      schema_context.vocabularies.contains(Known::JSON_Schema_Draft_6) ||
+      schema_context.vocabularies.contains(Known::JSON_Schema_Draft_7) ||
+      schema_context.vocabularies.contains(
+          Known::JSON_Schema_2019_09_Validation) ||
+      schema_context.vocabularies.contains(
+          Known::JSON_Schema_2020_12_Validation);
+  std::set<std::string> required;
+  if (imports_validation_vocabulary &&
+      schema_context.schema.defines("required") &&
+      schema_context.schema.at("required").is_array()) {
+    for (const auto &property :
+         schema_context.schema.at("required").as_array()) {
+      if (property.is_string() &&
+          // Only count the required property if its indeed in "properties"
+          properties.defines(property.to_string())) {
+        required.insert(property.to_string());
+      }
+    }
+  }
+
+  const auto &current_entry{static_frame_entry(context, schema_context)};
+  const auto inside_disjunctor{
+      is_inside_disjunctor(schema_context.relative_pointer) ||
+      // Check if any reference from `anyOf` or `oneOf` points to us
+      std::ranges::any_of(
+          context.frame.references(),
+          [&context, &current_entry](const auto &reference) {
+            if (!context.frame.locations().contains(
+                    {sourcemeta::core::SchemaReferenceType::Static,
+                     reference.second.destination})) {
+              return false;
+            }
+
+            const auto &target{
+                context.frame.locations()
+                    .at({sourcemeta::core::SchemaReferenceType::Static,
+                         reference.second.destination})
+                    .pointer};
+            return is_inside_disjunctor(reference.first.second) &&
+                   current_entry.pointer.initial() == target;
+          })};
+
+  if (!inside_disjunctor &&
+      schema_context.schema.defines("additionalProperties") &&
+      schema_context.schema.at("additionalProperties").is_boolean() &&
+      !schema_context.schema.at("additionalProperties").to_boolean() &&
+      // If all properties are required, we should still unroll
+      required.size() < size) {
+    return true;
+  }
+
+  return
+      // This strategy only makes sense if most of the properties are "optional"
+      required.size() <= (size / 4) &&
+      // If `properties` only defines a relatively small amount of properties,
+      // then its probably still faster to unroll
+      size > 5 &&
+      // Always unroll inside `oneOf` or `anyOf`, to have a
+      // better chance at quickly short-circuiting
+      (!inside_disjunctor ||
+       std::ranges::none_of(properties.as_object(), [&](const auto &pair) {
+         return pair.second.is_object() &&
+                ((imports_validation_vocabulary &&
+                  pair.second.defines("enum")) ||
+                 (imports_const && pair.second.defines("const")));
+       }));
+}
+
+auto is_integer_type_check(const Instruction &instruction) -> bool {
+  return (instruction.type == InstructionIndex::AssertionType ||
+          instruction.type == InstructionIndex::AssertionTypeStrict) &&
+         std::get<ValueType>(instruction.value) ==
+             sourcemeta::core::JSON::Type::Integer;
+}
+
+auto has_strict_integer_type(const Instructions &children) -> bool {
+  for (const auto &child : children) {
+    if (child.type == InstructionIndex::AssertionTypeStrict &&
+        std::get<ValueType>(child.value) ==
+            sourcemeta::core::JSON::Type::Integer) {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+auto is_integer_type_bounded_pattern(const Instructions &children) -> bool {
+  if (children.size() != 3) {
+    return false;
+  }
+
+  bool has_type{false};
+  bool has_min{false};
+  bool has_max{false};
+  for (const auto &child : children) {
+    if (is_integer_type_check(child)) {
+      has_type = true;
+    } else if (child.type == InstructionIndex::AssertionGreaterEqual &&
+               std::get<ValueJSON>(child.value).is_integer()) {
+      has_min = true;
+    } else if (child.type == InstructionIndex::AssertionLessEqual &&
+               std::get<ValueJSON>(child.value).is_integer()) {
+      has_max = true;
+    }
+  }
+
+  return has_type && has_min && has_max;
+}
+
+auto is_integer_type_lower_bound_pattern(const Instructions &children) -> bool {
+  if (children.size() != 2) {
+    return false;
+  }
+
+  bool has_type{false};
+  bool has_min{false};
+  for (const auto &child : children) {
+    if (is_integer_type_check(child)) {
+      has_type = true;
+    } else if (child.type == InstructionIndex::AssertionGreaterEqual &&
+               std::get<ValueJSON>(child.value).is_integer()) {
+      has_min = true;
+    }
+  }
+
+  return has_type && has_min;
+}
+
+auto extract_integer_lower_bound(const Instructions &children) -> std::int64_t {
+  for (const auto &child : children) {
+    if (child.type == InstructionIndex::AssertionGreaterEqual) {
+      return std::get<ValueJSON>(child.value).to_integer();
+    }
+  }
+
+  return 0;
+}
+
+auto extract_integer_bounds(const Instructions &children)
+    -> ValueIntegerBounds {
+  std::int64_t minimum{0};
+  std::int64_t maximum{0};
+  for (const auto &child : children) {
+    if (child.type == InstructionIndex::AssertionGreaterEqual) {
+      minimum = std::get<ValueJSON>(child.value).to_integer();
+    } else if (child.type == InstructionIndex::AssertionLessEqual) {
+      maximum = std::get<ValueJSON>(child.value).to_integer();
+    }
+  }
+
+  return {minimum, maximum};
+}
+
+auto compiler_draft3_applicator_properties_with_options(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &current,
+    const bool annotate, const bool track_evaluation) -> Instructions {
+  if (schema_context.is_property_name) {
+    return {};
+  }
+
+  if (!schema_context.schema.at(dynamic_context.keyword).is_object()) {
+    return {};
+  }
+
+  if (schema_context.schema.at(dynamic_context.keyword).empty()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "object") {
+    return {};
+  }
+
+  if (properties_as_loop(context, schema_context,
+                         schema_context.schema.at(dynamic_context.keyword))) {
+    ValueNamedIndexes indexes;
+    Instructions children;
+    std::size_t cursor = 0;
+
+    for (auto &&[name, substeps] : compile_properties(
+             context, schema_context, relative_dynamic_context(), current)) {
+      indexes.emplace(name, cursor);
+
+      if (track_evaluation) {
+        substeps.push_back(make(
+            sourcemeta::blaze::InstructionIndex::ControlEvaluate, context,
+            schema_context, relative_dynamic_context(), ValuePointer{name}));
+      }
+
+      if (annotate) {
+        substeps.push_back(
+            make(sourcemeta::blaze::InstructionIndex::AnnotationEmit, context,
+                 schema_context, relative_dynamic_context(),
+                 sourcemeta::core::JSON{name}));
+      }
+
+      // Note that the evaluator completely ignores this wrapper anyway
+      children.push_back(make(sourcemeta::blaze::InstructionIndex::ControlGroup,
+                              context, schema_context,
+                              relative_dynamic_context(), ValueNone{},
+                              std::move(substeps)));
+      cursor += 1;
+    }
+
+    if (context.mode == Mode::FastValidation && !track_evaluation &&
+        !schema_context.schema.defines("patternProperties") &&
+        schema_context.schema.defines("additionalProperties") &&
+        schema_context.schema.at("additionalProperties").is_boolean() &&
+        !schema_context.schema.at("additionalProperties").to_boolean()) {
+      return {
+          make(sourcemeta::blaze::InstructionIndex::LoopPropertiesMatchClosed,
+               context, schema_context, dynamic_context, std::move(indexes),
+               std::move(children))};
+    }
+
+    return {make(sourcemeta::blaze::InstructionIndex::LoopPropertiesMatch,
+                 context, schema_context, dynamic_context, std::move(indexes),
+                 std::move(children))};
+  }
+
+  Instructions children;
+
+  const auto effective_dynamic_context{context.mode == Mode::FastValidation
+                                           ? dynamic_context
+                                           : relative_dynamic_context()};
+
+  const auto assume_object{schema_context.schema.defines("type") &&
+                           schema_context.schema.at("type").is_string() &&
+                           schema_context.schema.at("type").to_string() ==
+                               "object"};
+
+  auto properties{compile_properties(context, schema_context,
+                                     effective_dynamic_context, current)};
+
+  if (context.mode == Mode::FastValidation &&
+      schema_context.schema.defines("additionalProperties") &&
+      schema_context.schema.at("additionalProperties").is_boolean() &&
+      !schema_context.schema.at("additionalProperties").to_boolean() &&
+      // TODO: Check that the validation vocabulary is present
+      schema_context.schema.defines("required") &&
+      schema_context.schema.at("required").is_array() &&
+      schema_context.schema.at("required").size() ==
+          schema_context.schema.at(dynamic_context.keyword).size() &&
+      std::ranges::all_of(properties, [&schema_context](const auto &property) {
+        return schema_context.schema.at("required")
+            .contains(sourcemeta::core::JSON{property.first});
+      })) {
+    if (std::ranges::all_of(properties, [](const auto &property) {
+          return property.second.size() == 1 &&
+                 property.second.front().type ==
+                     InstructionIndex::AssertionTypeStrict;
+        })) {
+      std::set<ValueType> types;
+      for (const auto &property : properties) {
+        types.insert(std::get<ValueType>(property.second.front().value));
+      }
+
+      if (types.size() == 1 &&
+          !schema_context.schema.defines("patternProperties")) {
+        if (schema_context.schema.defines("required") && assume_object) {
+          auto required_copy = schema_context.schema.at("required");
+          std::ranges::sort(required_copy.as_array());
+          ValueStringSet required{json_array_to_string_set(required_copy)};
+          if (is_closed_properties_required(schema_context.schema, required)) {
+            sourcemeta::core::PropertyHashJSON<ValueString> hasher;
+            std::vector<std::pair<ValueString, ValueStringSet::hash_type>>
+                perfect_hashes;
+            for (const auto &entry : required) {
+              assert(required.contains(entry.first, entry.second));
+              if (hasher.is_perfect(entry.second)) {
+                perfect_hashes.emplace_back(entry.first, entry.second);
+              }
+            }
+
+            if (perfect_hashes.size() == required.size()) {
+              return {make(sourcemeta::blaze::InstructionIndex::
+                               LoopPropertiesExactlyTypeStrictHash,
+                           context, schema_context, dynamic_context,
+                           ValueTypedHashes{*types.cbegin(),
+                                            to_string_hashes(perfect_hashes)})};
+            }
+
+            return {make(
+                sourcemeta::blaze::InstructionIndex::
+                    LoopPropertiesExactlyTypeStrict,
+                context, schema_context, dynamic_context,
+                ValueTypedProperties{*types.cbegin(), std::move(required)})};
+          }
+        }
+
+        return {
+            make(sourcemeta::blaze::InstructionIndex::LoopPropertiesTypeStrict,
+                 context, schema_context, dynamic_context, *types.cbegin())};
+      }
+    }
+
+    if (std::ranges::all_of(properties, [](const auto &property) {
+          return property.second.size() == 1 &&
+                 property.second.front().type ==
+                     InstructionIndex::AssertionType;
+        })) {
+      std::set<ValueType> types;
+      for (const auto &property : properties) {
+        types.insert(std::get<ValueType>(property.second.front().value));
+      }
+
+      if (types.size() == 1) {
+        return {make(sourcemeta::blaze::InstructionIndex::LoopPropertiesType,
+                     context, schema_context, dynamic_context,
+                     *types.cbegin())};
+      }
+    }
+  }
+
+  auto attempt_object_fusion{context.mode == Mode::FastValidation &&
+                             !annotate && !track_evaluation && assume_object};
+  if (attempt_object_fusion) {
+    for (const auto &entry : schema_context.schema.as_object()) {
+      const auto &keyword{entry.first};
+      if (keyword == "type" || keyword == "required" ||
+          keyword == dynamic_context.keyword) {
+        continue;
+      }
+
+      if (keyword == "additionalProperties" && entry.second.is_boolean() &&
+          entry.second.to_boolean()) {
+        continue;
+      }
+
+      const auto &keyword_type{
+          context.walker(keyword, schema_context.vocabularies).type};
+      using enum sourcemeta::core::SchemaKeywordType;
+      if (keyword_type == Assertion || keyword_type == Annotation ||
+          keyword_type == Unknown || keyword_type == Comment ||
+          keyword_type == Other || keyword_type == LocationMembers) {
+        continue;
+      }
+
+      attempt_object_fusion = false;
+      break;
+    }
+  }
+  ValueObjectProperties fusion_entries;
+  Instructions fusion_children;
+  bool fusion_possible{attempt_object_fusion};
+
+  for (auto &&[name, substeps] : properties) {
+    if (annotate) {
+      substeps.push_back(
+          make(sourcemeta::blaze::InstructionIndex::AnnotationEmit, context,
+               schema_context, effective_dynamic_context,
+               sourcemeta::core::JSON{name}));
+    }
+
+    // Optimize `properties` where its subschemas just include a type check
+
+    if (context.mode == Mode::FastValidation && track_evaluation &&
+        substeps.size() == 1 &&
+        substeps.front().type == InstructionIndex::AssertionTypeStrict) {
+      children.push_back(rephrase(context,
+                                  sourcemeta::blaze::InstructionIndex::
+                                      AssertionPropertyTypeStrictEvaluate,
+                                  substeps.front()));
+    } else if (context.mode == Mode::FastValidation && track_evaluation &&
+               substeps.size() == 1 &&
+               substeps.front().type == InstructionIndex::AssertionType) {
+      children.push_back(rephrase(
+          context,
+          sourcemeta::blaze::InstructionIndex::AssertionPropertyTypeEvaluate,
+          substeps.front()));
+    } else if (context.mode == Mode::FastValidation && track_evaluation &&
+               substeps.size() == 1 &&
+               substeps.front().type ==
+                   InstructionIndex::AssertionTypeStrictAny) {
+      children.push_back(rephrase(context,
+                                  sourcemeta::blaze::InstructionIndex::
+                                      AssertionPropertyTypeStrictAnyEvaluate,
+                                  substeps.front()));
+
+      // NOLINTBEGIN(bugprone-branch-clone)
+    } else if (!fusion_possible && context.mode == Mode::FastValidation &&
+               substeps.size() == 1 &&
+               substeps.front().type ==
+                   InstructionIndex::AssertionPropertyTypeStrict) {
+      children.push_back(
+          unroll(context, substeps.front(),
+                 effective_dynamic_context.base_instance_location));
+    } else if (!fusion_possible && context.mode == Mode::FastValidation &&
+               substeps.size() == 1 &&
+               substeps.front().type ==
+                   InstructionIndex::AssertionPropertyType) {
+      children.push_back(
+          unroll(context, substeps.front(),
+                 effective_dynamic_context.base_instance_location));
+    } else if (!fusion_possible && context.mode == Mode::FastValidation &&
+               substeps.size() == 1 &&
+               substeps.front().type ==
+                   InstructionIndex::AssertionPropertyTypeStrictAny) {
+      children.push_back(
+          unroll(context, substeps.front(),
+                 effective_dynamic_context.base_instance_location));
+      // NOLINTEND(bugprone-branch-clone)
+
+    } else {
+      if (track_evaluation) {
+        auto new_base_instance_location{
+            effective_dynamic_context.base_instance_location};
+        new_base_instance_location.push_back({name});
+        substeps.push_back(
+            make(sourcemeta::blaze::InstructionIndex::Evaluate, context,
+                 schema_context,
+                 DynamicContext{
+                     .keyword = effective_dynamic_context.keyword,
+                     .base_schema_location =
+                         effective_dynamic_context.base_schema_location,
+                     .base_instance_location = new_base_instance_location},
+                 ValueNone{}));
+      }
+
+      if (context.mode == Mode::FastValidation && !substeps.empty()) {
+        if (is_integer_type_bounded_pattern(substeps)) {
+          auto bounds = extract_integer_bounds(substeps);
+          const auto index =
+              has_strict_integer_type(substeps)
+                  ? InstructionIndex::AssertionTypeIntegerBoundedStrict
+                  : InstructionIndex::AssertionTypeIntegerBounded;
+          auto instance_location = substeps.front().relative_instance_location;
+          substeps.clear();
+          auto fused = make(index, context, schema_context,
+                            relative_dynamic_context(), bounds);
+          fused.relative_instance_location = std::move(instance_location);
+          substeps.push_back(std::move(fused));
+        } else if (is_integer_type_lower_bound_pattern(substeps)) {
+          const auto minimum = extract_integer_lower_bound(substeps);
+          const auto index =
+              has_strict_integer_type(substeps)
+                  ? InstructionIndex::AssertionTypeIntegerLowerBoundStrict
+                  : InstructionIndex::AssertionTypeIntegerLowerBound;
+          auto instance_location = substeps.front().relative_instance_location;
+          substeps.clear();
+          auto fused =
+              make(index, context, schema_context, relative_dynamic_context(),
+                   ValueIntegerBounds{minimum, 0});
+          fused.relative_instance_location = std::move(instance_location);
+          substeps.push_back(std::move(fused));
+        } else if (substeps.size() == 2) {
+          bool has_items_bounded{false};
+          bool has_array_type{false};
+          std::size_t items_index{0};
+          std::size_t array_index{0};
+          for (std::size_t step_index = 0; step_index < 2; step_index++) {
+            if (substeps[step_index].type ==
+                InstructionIndex::LoopItemsIntegerBounded) {
+              has_items_bounded = true;
+              items_index = step_index;
+            } else if (substeps[step_index].type ==
+                       InstructionIndex::AssertionTypeArrayBounded) {
+              has_array_type = true;
+              array_index = step_index;
+            }
+          }
+
+          if (has_items_bounded && has_array_type) {
+            auto integer_bounds{
+                std::get<ValueIntegerBounds>(substeps[items_index].value)};
+            auto range{std::get<ValueRange>(substeps[array_index].value)};
+            auto instance_location =
+                substeps[items_index].relative_instance_location;
+            Value fused_value{
+                ValueIntegerBoundsWithSize{integer_bounds, std::move(range)}};
+            substeps.clear();
+            auto fused =
+                make(InstructionIndex::LoopItemsIntegerBoundedSized, context,
+                     schema_context, effective_dynamic_context, fused_value);
+            fused.relative_instance_location = std::move(instance_location);
+            substeps.push_back(std::move(fused));
+          }
+        }
+      }
+
+      if (fusion_possible && substeps.size() >= 2 &&
+          std::ranges::any_of(substeps, [](const auto &step) {
+            return step.type ==
+                   InstructionIndex::AssertionObjectPropertiesSimple;
+          })) {
+        std::erase_if(substeps, [](const auto &step) {
+          if (step.type == InstructionIndex::AssertionDefinesAllStrict ||
+              step.type == InstructionIndex::AssertionDefinesAll) {
+            return true;
+          }
+
+          if ((step.type == InstructionIndex::AssertionTypeStrict ||
+               step.type == InstructionIndex::AssertionType) &&
+              std::get<ValueType>(step.value) ==
+                  sourcemeta::core::JSON::Type::Object) {
+            return true;
+          }
+
+          return false;
+        });
+      }
+
+      if (fusion_possible && substeps.size() == 1 &&
+          substeps.front().type != InstructionIndex::ControlJump &&
+          substeps.front().type != InstructionIndex::ControlDynamicAnchorJump) {
+        const auto is_required{
+            assume_object && schema_context.schema.defines("required") &&
+            schema_context.schema.at("required").is_array() &&
+            schema_context.schema.at("required")
+                .contains(sourcemeta::core::JSON{name})};
+        auto prop{make_property(name)};
+        auto fusion_child{substeps.front()};
+        fusion_child.relative_instance_location = {};
+        auto fusion_extra{context.extra[fusion_child.extra_index]};
+        fusion_extra.relative_schema_location = {};
+        fusion_child.extra_index = context.extra.size();
+        context.extra.push_back(std::move(fusion_extra));
+
+        fusion_entries.emplace_back(prop.first, prop.second, is_required);
+        fusion_children.push_back(std::move(fusion_child));
+      } else {
+        fusion_possible = false;
+      }
+
+      if (!substeps.empty()) {
+        // As a performance shortcut
+        if (effective_dynamic_context.base_instance_location.empty()) {
+          if (assume_object &&
+              // TODO: Check that the validation vocabulary is present
+              schema_context.schema.defines("required") &&
+              schema_context.schema.at("required").is_array() &&
+              schema_context.schema.at("required")
+                  .contains(sourcemeta::core::JSON{name})) {
+            for (auto &&step : substeps) {
+              children.push_back(std::move(step));
+            }
+          } else {
+            children.push_back(make(sourcemeta::blaze::InstructionIndex::
+                                        ControlGroupWhenDefinesDirect,
+                                    context, schema_context,
+                                    effective_dynamic_context,
+                                    make_property(name), std::move(substeps)));
+          }
+        } else {
+          children.push_back(
+              make(sourcemeta::blaze::InstructionIndex::ControlGroupWhenDefines,
+                   context, schema_context, effective_dynamic_context,
+                   make_property(name), std::move(substeps)));
+        }
+      }
+    }
+  }
+
+  if (context.mode == Mode::FastValidation) {
+    if (fusion_possible && !fusion_entries.empty()) {
+      if (schema_context.schema.defines("required") &&
+          schema_context.schema.at("required").is_array()) {
+        for (const auto &req :
+             schema_context.schema.at("required").as_array()) {
+          if (!req.is_string()) {
+            continue;
+          }
+          const auto &req_name{req.to_string()};
+          bool already_tracked{false};
+          for (const auto &entry : fusion_entries) {
+            if (std::get<0>(entry) == req_name) {
+              already_tracked = true;
+              break;
+            }
+          }
+          if (!already_tracked) {
+            auto prop{make_property(req_name)};
+            fusion_entries.emplace_back(prop.first, prop.second, true);
+          }
+        }
+      }
+
+      if (fusion_entries.size() > 32) {
+        return children;
+      }
+
+      return {make(InstructionIndex::AssertionObjectPropertiesSimple, context,
+                   schema_context, dynamic_context,
+                   Value{std::move(fusion_entries)},
+                   std::move(fusion_children))};
+    }
+
+    return children;
+  } else if (children.empty()) {
+    return {};
+  } else {
+    return {make(sourcemeta::blaze::InstructionIndex::LogicalWhenType, context,
+                 schema_context, dynamic_context,
+                 sourcemeta::core::JSON::Type::Object, std::move(children))};
+  }
+}
+
+auto compiler_draft3_applicator_properties(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &current)
+    -> Instructions {
+  return compiler_draft3_applicator_properties_with_options(
+      context, schema_context, dynamic_context, current, false, false);
+}
+
+auto compiler_draft3_applicator_patternproperties_with_options(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const bool annotate,
+    const bool track_evaluation) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_object()) {
+    return {};
+  }
+
+  if (schema_context.schema.at(dynamic_context.keyword).empty()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "object") {
+    return {};
+  }
+
+  Instructions children;
+
+  // To guarantee ordering
+  std::vector<std::string> patterns;
+  for (auto &entry :
+       schema_context.schema.at(dynamic_context.keyword).as_object()) {
+    patterns.push_back(entry.first);
+  }
+
+  std::ranges::sort(patterns);
+
+  // For each regular expression and corresponding subschema in the object
+  for (const auto &pattern : patterns) {
+    auto substeps{compile(context, schema_context, relative_dynamic_context(),
+                          sourcemeta::blaze::make_weak_pointer(pattern))};
+
+    if (annotate) {
+      substeps.push_back(make(
+          sourcemeta::blaze::InstructionIndex::AnnotationBasenameToParent,
+          context, schema_context, relative_dynamic_context(), ValueNone{}));
+    }
+
+    if (track_evaluation) {
+      substeps.push_back(
+          make(sourcemeta::blaze::InstructionIndex::ControlEvaluate, context,
+               schema_context, relative_dynamic_context(), ValuePointer{}));
+    }
+
+    if (context.mode == Mode::FastValidation && !track_evaluation &&
+        patterns.size() == 1 &&
+        (!schema_context.schema.defines("properties") ||
+         (schema_context.schema.at("properties").is_object() &&
+          schema_context.schema.at("properties").empty())) &&
+        schema_context.schema.defines("additionalProperties") &&
+        schema_context.schema.at("additionalProperties").is_boolean() &&
+        !schema_context.schema.at("additionalProperties").to_boolean()) {
+      children.push_back(
+          make(sourcemeta::blaze::InstructionIndex::LoopPropertiesRegexClosed,
+               context, schema_context, dynamic_context,
+               ValueRegex{.first = parse_regex(pattern, schema_context.base,
+                                               schema_context.relative_pointer),
+                          .second = pattern},
+               std::move(substeps)));
+
+      // If the `patternProperties` subschema for the given pattern does
+      // nothing, then we can avoid generating an entire loop for it
+    } else if (!substeps.empty()) {
+      const auto maybe_prefix{pattern_as_prefix(pattern)};
+      if (maybe_prefix.has_value()) {
+        children.push_back(
+            make(sourcemeta::blaze::InstructionIndex::LoopPropertiesStartsWith,
+                 context, schema_context, dynamic_context,
+                 ValueString{maybe_prefix.value()}, std::move(substeps)));
+      } else {
+        children.push_back(make(
+            sourcemeta::blaze::InstructionIndex::LoopPropertiesRegex, context,
+            schema_context, dynamic_context,
+            ValueRegex{.first = parse_regex(pattern, schema_context.base,
+                                            schema_context.relative_pointer),
+                       .second = pattern},
+            std::move(substeps)));
+      }
+    }
+  }
+
+  return children;
+}
+
+auto compiler_draft3_applicator_patternproperties(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &)
+    -> Instructions {
+  return compiler_draft3_applicator_patternproperties_with_options(
+      context, schema_context, dynamic_context, false, false);
+}
+
+auto compiler_draft3_applicator_additionalproperties_with_options(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const bool annotate,
+    const bool track_evaluation) -> Instructions {
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "object") {
+    return {};
+  }
+
+  Instructions children{compile(context, schema_context,
+                                relative_dynamic_context(),
+                                sourcemeta::core::empty_weak_pointer,
+                                sourcemeta::core::empty_weak_pointer)};
+
+  if (annotate) {
+    children.push_back(
+        make(sourcemeta::blaze::InstructionIndex::AnnotationBasenameToParent,
+             context, schema_context, relative_dynamic_context(), ValueNone{}));
+  }
+
+  ValueStringSet filter_strings;
+  ValueStrings filter_prefixes;
+  std::vector<ValueRegex> filter_regexes;
+
+  if (schema_context.schema.defines("properties") &&
+      schema_context.schema.at("properties").is_object()) {
+    for (const auto &entry :
+         schema_context.schema.at("properties").as_object()) {
+      filter_strings.insert(entry.first);
+    }
+  }
+
+  if (schema_context.schema.defines("patternProperties") &&
+      schema_context.schema.at("patternProperties").is_object()) {
+    for (const auto &entry :
+         schema_context.schema.at("patternProperties").as_object()) {
+      const auto maybe_prefix{pattern_as_prefix(entry.first)};
+      if (maybe_prefix.has_value()) {
+        filter_prefixes.push_back(maybe_prefix.value());
+      } else {
+        static const std::string pattern_properties_keyword{
+            "patternProperties"};
+        filter_regexes.push_back(
+            {parse_regex(entry.first, schema_context.base,
+                         schema_context.relative_pointer.initial().concat(
+                             sourcemeta::blaze::make_weak_pointer(
+                                 pattern_properties_keyword))),
+             entry.first});
+      }
+    }
+  }
+
+  // For performance, if a schema sets `additionalProperties: true` (or its
+  // variants), we don't need to do anything
+  if (!track_evaluation && children.empty()) {
+    return {};
+  }
+
+  // When `additionalProperties: false` with only `properties` (no
+  // patternProperties), and `properties` is compiled as a loop
+  // (LoopPropertiesMatchClosed), that loop already handles rejecting unknown
+  // properties, so we don't need to emit anything for `additionalProperties`
+  if (context.mode == Mode::FastValidation && children.size() == 1 &&
+      children.front().type == InstructionIndex::AssertionFail &&
+      !filter_strings.empty() && filter_prefixes.empty() &&
+      filter_regexes.empty() &&
+      properties_as_loop(context, schema_context,
+                         schema_context.schema.at("properties"))) {
+    return {};
+  }
+
+  // When all properties are required and `additionalProperties: false`,
+  // the `required` keyword compiles to `AssertionDefinesExactly` which already
+  // checks that the object has exactly the required properties, so we don't
+  // need to emit anything for `additionalProperties`
+  if (context.mode == Mode::FastValidation && children.size() == 1 &&
+      children.front().type == InstructionIndex::AssertionFail &&
+      !filter_strings.empty() && filter_prefixes.empty() &&
+      filter_regexes.empty() && schema_context.schema.defines("required") &&
+      schema_context.schema.at("required").is_array() &&
+      is_closed_properties_required(
+          schema_context.schema,
+          json_array_to_string_set(schema_context.schema.at("required")))) {
+    return {};
+  }
+
+  if (context.mode == Mode::FastValidation && filter_strings.empty() &&
+      filter_prefixes.empty() && filter_regexes.size() == 1 &&
+      !track_evaluation && !children.empty() &&
+      children.front().type == InstructionIndex::AssertionFail) {
+    return {};
+  }
+
+  if (!filter_strings.empty() || !filter_prefixes.empty() ||
+      !filter_regexes.empty()) {
+    if (track_evaluation) {
+      children.push_back(
+          make(sourcemeta::blaze::InstructionIndex::ControlEvaluate, context,
+               schema_context, relative_dynamic_context(), ValuePointer{}));
+    }
+
+    return {make(sourcemeta::blaze::InstructionIndex::LoopPropertiesExcept,
+                 context, schema_context, dynamic_context,
+                 ValuePropertyFilter{std::move(filter_strings),
+                                     std::move(filter_prefixes),
+                                     std::move(filter_regexes)},
+                 std::move(children))};
+  } else if (track_evaluation) {
+    if (children.empty()) {
+      return {make(sourcemeta::blaze::InstructionIndex::Evaluate, context,
+                   schema_context, dynamic_context, ValueNone{})};
+    }
+
+    return {make(sourcemeta::blaze::InstructionIndex::LoopPropertiesEvaluate,
+                 context, schema_context, dynamic_context, ValueNone{},
+                 std::move(children))};
+  } else if (children.size() == 1 &&
+             children.front().type == InstructionIndex::AssertionFail) {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionObjectSizeLess,
+                 context, schema_context, dynamic_context,
+                 ValueUnsignedInteger{1})};
+  } else {
+    return {make(sourcemeta::blaze::InstructionIndex::LoopProperties, context,
+                 schema_context, dynamic_context, ValueNone{},
+                 std::move(children))};
+  }
+}
+
+auto compiler_draft3_applicator_additionalproperties(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &)
+    -> Instructions {
+  return compiler_draft3_applicator_additionalproperties_with_options(
+      context, schema_context, dynamic_context, false, false);
+}
+
+auto compiler_draft3_validation_pattern(const Context &context,
+                                        const SchemaContext &schema_context,
+                                        const DynamicContext &dynamic_context,
+                                        const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_string()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "string") {
+    return {};
+  }
+
+  const auto &regex_string{
+      schema_context.schema.at(dynamic_context.keyword).to_string()};
+  return {
+      make(sourcemeta::blaze::InstructionIndex::AssertionRegex, context,
+           schema_context, dynamic_context,
+           ValueRegex{.first = parse_regex(regex_string, schema_context.base,
+                                           schema_context.relative_pointer),
+                      .second = regex_string})};
+}
+
+auto compiler_draft3_applicator_items_array(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const bool annotate,
+    const bool track_evaluation) -> Instructions {
+  if (schema_context.is_property_name) {
+    return {};
+  }
+
+  if (!schema_context.schema.at(dynamic_context.keyword).is_array()) {
+    return {};
+  }
+
+  const auto items_size{
+      schema_context.schema.at(dynamic_context.keyword).size()};
+  if (items_size == 0) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  // Precompile subschemas
+  std::vector<Instructions> subschemas;
+  subschemas.reserve(items_size);
+  const auto &array{
+      schema_context.schema.at(dynamic_context.keyword).as_array()};
+  for (auto iterator{array.cbegin()}; iterator != array.cend(); ++iterator) {
+    subschemas.push_back(compile(context, schema_context,
+                                 relative_dynamic_context(),
+                                 {subschemas.size()}, {subschemas.size()}));
+  }
+
+  Instructions children;
+  for (std::size_t cursor = 0; cursor < items_size; cursor++) {
+    Instructions subchildren;
+    for (std::size_t index = 0; index < cursor + 1; index++) {
+      for (const auto &substep : subschemas.at(index)) {
+        subchildren.push_back(substep);
+      }
+    }
+
+    if (annotate) {
+      subchildren.push_back(
+          make(sourcemeta::blaze::InstructionIndex::AnnotationEmit, context,
+               schema_context, relative_dynamic_context(),
+               sourcemeta::core::JSON{cursor}));
+    }
+
+    children.push_back(make(sourcemeta::blaze::InstructionIndex::ControlGroup,
+                            context, schema_context, relative_dynamic_context(),
+                            ValueNone{}, std::move(subchildren)));
+  }
+
+  Instructions tail;
+  for (const auto &subschema : subschemas) {
+    for (const auto &substep : subschema) {
+      tail.push_back(substep);
+    }
+  }
+
+  if (annotate) {
+    tail.push_back(make(sourcemeta::blaze::InstructionIndex::AnnotationEmit,
+                        context, schema_context, relative_dynamic_context(),
+                        sourcemeta::core::JSON{children.size() - 1}));
+    tail.push_back(make(sourcemeta::blaze::InstructionIndex::AnnotationEmit,
+                        context, schema_context, relative_dynamic_context(),
+                        sourcemeta::core::JSON{true}));
+  }
+
+  children.push_back(make(sourcemeta::blaze::InstructionIndex::ControlGroup,
+                          context, schema_context, relative_dynamic_context(),
+                          ValueNone{}, std::move(tail)));
+
+  if (track_evaluation) {
+    return {
+        make(sourcemeta::blaze::InstructionIndex::AssertionArrayPrefixEvaluate,
+             context, schema_context, dynamic_context, ValueNone{},
+             std::move(children))};
+  } else {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionArrayPrefix,
+                 context, schema_context, dynamic_context, ValueNone{},
+                 std::move(children))};
+  }
+}
+
+auto is_number_type_check(const Instruction &instruction) -> bool {
+  if (instruction.type != InstructionIndex::AssertionTypeStrictAny) {
+    return false;
+  }
+
+  const auto &value{std::get<ValueTypes>(instruction.value)};
+  const auto numeric_count{
+      static_cast<std::size_t>(value.test(
+          std::to_underlying(sourcemeta::core::JSON::Type::Integer))) +
+      static_cast<std::size_t>(
+          value.test(std::to_underlying(sourcemeta::core::JSON::Type::Real))) +
+      static_cast<std::size_t>(value.test(
+          std::to_underlying(sourcemeta::core::JSON::Type::Decimal)))};
+  return numeric_count >= 2 && value.count() == numeric_count;
+}
+
+auto is_integer_bounded_pattern(const Instructions &children) -> bool {
+  if (children.size() != 3) {
+    return false;
+  }
+
+  bool has_type{false};
+  bool has_min{false};
+  bool has_max{false};
+  for (const auto &child : children) {
+    if (is_number_type_check(child)) {
+      has_type = true;
+    } else if (child.type == InstructionIndex::AssertionGreaterEqual) {
+      if (!std::get<ValueJSON>(child.value).is_integer()) {
+        return false;
+      }
+      has_min = true;
+    } else if (child.type == InstructionIndex::AssertionLessEqual) {
+      if (!std::get<ValueJSON>(child.value).is_integer()) {
+        return false;
+      }
+      has_max = true;
+    }
+  }
+
+  return has_type && has_min && has_max;
+}
+
+auto compiler_draft3_applicator_items_with_options(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const bool annotate,
+    const bool track_evaluation) -> Instructions {
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  if (is_schema(schema_context.schema.at(dynamic_context.keyword))) {
+    if (annotate || track_evaluation) {
+      Instructions subchildren{compile(context, schema_context,
+                                       relative_dynamic_context(),
+                                       sourcemeta::core::empty_weak_pointer,
+                                       sourcemeta::core::empty_weak_pointer)};
+
+      Instructions children;
+
+      if (!subchildren.empty()) {
+        children.push_back(make(sourcemeta::blaze::InstructionIndex::LoopItems,
+                                context, schema_context, dynamic_context,
+                                ValueNone{}, std::move(subchildren)));
+      }
+
+      if (!annotate && !track_evaluation) {
+        return children;
+      }
+
+      Instructions tail;
+
+      if (annotate) {
+        tail.push_back(make(sourcemeta::blaze::InstructionIndex::AnnotationEmit,
+                            context, schema_context, relative_dynamic_context(),
+                            sourcemeta::core::JSON{true}));
+      }
+
+      if (track_evaluation) {
+        tail.push_back(
+            make(sourcemeta::blaze::InstructionIndex::ControlEvaluate, context,
+                 schema_context, relative_dynamic_context(), ValuePointer{}));
+      }
+
+      children.push_back(
+          make(sourcemeta::blaze::InstructionIndex::LogicalWhenType, context,
+               schema_context, dynamic_context,
+               sourcemeta::core::JSON::Type::Array, std::move(tail)));
+
+      return children;
+    }
+
+    Instructions children{compile(context, schema_context,
+                                  relative_dynamic_context(),
+                                  sourcemeta::core::empty_weak_pointer,
+                                  sourcemeta::core::empty_weak_pointer)};
+    if (track_evaluation) {
+      children.push_back(
+          make(sourcemeta::blaze::InstructionIndex::ControlEvaluate, context,
+               schema_context, relative_dynamic_context(), ValuePointer{}));
+    }
+
+    if (children.empty()) {
+      return {};
+    }
+
+    if (context.mode == Mode::FastValidation && children.size() == 3 &&
+        is_integer_bounded_pattern(children)) {
+      return {make(sourcemeta::blaze::InstructionIndex::LoopItemsIntegerBounded,
+                   context, schema_context, dynamic_context,
+                   extract_integer_bounds(children))};
+    }
+
+    if (context.mode == Mode::FastValidation && children.size() == 1) {
+      if (children.front().type == InstructionIndex::AssertionTypeStrict) {
+        return {make(sourcemeta::blaze::InstructionIndex::LoopItemsTypeStrict,
+                     context, schema_context, dynamic_context,
+                     children.front().value)};
+      } else if (children.front().type == InstructionIndex::AssertionType) {
+        return {make(sourcemeta::blaze::InstructionIndex::LoopItemsType,
+                     context, schema_context, dynamic_context,
+                     children.front().value)};
+      } else if (children.front().type ==
+                 InstructionIndex::AssertionTypeStrictAny) {
+        return {make(
+            sourcemeta::blaze::InstructionIndex::LoopItemsTypeStrictAny,
+            context, schema_context, dynamic_context, children.front().value)};
+      } else if (children.front().type ==
+                 InstructionIndex::LoopPropertiesExactlyTypeStrictHash) {
+        auto value_copy = children.front().value;
+        auto current{make(sourcemeta::blaze::InstructionIndex::LoopItems,
+                          context, schema_context, dynamic_context, ValueNone{},
+                          std::move(children))};
+        if (std::get<ValueTypedHashes>(value_copy).second.first.size() == 3) {
+          return {Instruction{.type = sourcemeta::blaze::InstructionIndex::
+                                  LoopItemsPropertiesExactlyTypeStrictHash3,
+                              .relative_instance_location =
+                                  current.relative_instance_location,
+                              .value = std::move(value_copy),
+                              .children = {},
+                              .extra_index = current.extra_index}};
+        }
+
+        return {Instruction{.type = sourcemeta::blaze::InstructionIndex::
+                                LoopItemsPropertiesExactlyTypeStrictHash,
+                            .relative_instance_location =
+                                current.relative_instance_location,
+                            .value = std::move(value_copy),
+                            .children = {},
+                            .extra_index = current.extra_index}};
+      }
+    }
+
+    return {make(sourcemeta::blaze::InstructionIndex::LoopItems, context,
+                 schema_context, dynamic_context, ValueNone{},
+                 std::move(children))};
+  }
+
+  return compiler_draft3_applicator_items_array(
+      context, schema_context, dynamic_context, annotate, track_evaluation);
+}
+
+auto compiler_draft3_applicator_items(const Context &context,
+                                      const SchemaContext &schema_context,
+                                      const DynamicContext &dynamic_context,
+                                      const Instructions &) -> Instructions {
+  return compiler_draft3_applicator_items_with_options(
+      context, schema_context, dynamic_context, false, false);
+}
+
+auto compiler_draft3_applicator_additionalitems_from_cursor(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const std::size_t cursor,
+    const bool annotate, const bool track_evaluation) -> Instructions {
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  Instructions subchildren{compile(context, schema_context,
+                                   relative_dynamic_context(),
+                                   sourcemeta::core::empty_weak_pointer,
+                                   sourcemeta::core::empty_weak_pointer)};
+
+  Instructions children;
+
+  if (!subchildren.empty()) {
+    if (context.mode == Mode::FastValidation && cursor == 0 && !annotate &&
+        !track_evaluation && is_integer_bounded_pattern(subchildren)) {
+      children.push_back(
+          make(sourcemeta::blaze::InstructionIndex::LoopItemsIntegerBounded,
+               context, schema_context, dynamic_context,
+               extract_integer_bounds(subchildren)));
+      return children;
+    }
+
+    children.push_back(make(sourcemeta::blaze::InstructionIndex::LoopItemsFrom,
+                            context, schema_context, dynamic_context,
+                            ValueUnsignedInteger{cursor},
+                            std::move(subchildren)));
+  }
+
+  // Avoid one extra wrapper instruction if possible
+  if (!annotate && !track_evaluation) {
+    return children;
+  }
+
+  Instructions tail;
+
+  if (annotate) {
+    tail.push_back(make(sourcemeta::blaze::InstructionIndex::AnnotationEmit,
+                        context, schema_context, relative_dynamic_context(),
+                        sourcemeta::core::JSON{true}));
+  }
+
+  if (track_evaluation) {
+    tail.push_back(make(sourcemeta::blaze::InstructionIndex::ControlEvaluate,
+                        context, schema_context, relative_dynamic_context(),
+                        ValuePointer{}));
+  }
+
+  assert(!tail.empty());
+  children.push_back(
+      make(sourcemeta::blaze::InstructionIndex::LogicalWhenArraySizeGreater,
+           context, schema_context, dynamic_context,
+           ValueUnsignedInteger{cursor}, std::move(tail)));
+
+  return children;
+}
+
+auto compiler_draft3_applicator_additionalitems_with_options(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const bool annotate,
+    const bool track_evaluation) -> Instructions {
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  assert(schema_context.schema.is_object());
+
+  // Nothing to do here
+  if (!schema_context.schema.defines("items") ||
+      schema_context.schema.at("items").is_object()) {
+    return {};
+  }
+
+  const auto cursor{(schema_context.schema.defines("items") &&
+                     schema_context.schema.at("items").is_array())
+                        ? schema_context.schema.at("items").size()
+                        : 0};
+
+  return compiler_draft3_applicator_additionalitems_from_cursor(
+      context, schema_context, dynamic_context, cursor, annotate,
+      track_evaluation);
+}
+
+auto compiler_draft3_applicator_additionalitems(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &)
+    -> Instructions {
+  return compiler_draft3_applicator_additionalitems_with_options(
+      context, schema_context, dynamic_context, false, false);
+}
+
+auto compiler_draft3_validation_enum(const Context &context,
+                                     const SchemaContext &schema_context,
+                                     const DynamicContext &dynamic_context,
+                                     const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_array()) {
+    return {};
+  }
+
+  if (schema_context.schema.at(dynamic_context.keyword).empty()) {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionFail, context,
+                 schema_context, dynamic_context, ValueNone{})};
+  }
+
+  if (schema_context.schema.at(dynamic_context.keyword).size() == 1) {
+    return {
+        make(sourcemeta::blaze::InstructionIndex::AssertionEqual, context,
+             schema_context, dynamic_context,
+             sourcemeta::core::JSON{
+                 schema_context.schema.at(dynamic_context.keyword).front()})};
+  }
+
+  std::vector<std::pair<sourcemeta::blaze::ValueString,
+                        sourcemeta::blaze::ValueStringSet::hash_type>>
+      perfect_string_hashes;
+  ValueSet options;
+  sourcemeta::core::PropertyHashJSON<ValueString> hasher;
+  for (const auto &option :
+       schema_context.schema.at(dynamic_context.keyword).as_array()) {
+    if (option.is_string()) {
+      const auto hash{hasher(option.to_string())};
+      if (hasher.is_perfect(hash)) {
+        perfect_string_hashes.emplace_back(option.to_string(), hash);
+      }
+    }
+
+    options.insert(option);
+  }
+
+  // Only apply this optimisation on fast validation, as it
+  // can affect error messages
+  if (context.mode == Mode::FastValidation &&
+      perfect_string_hashes.size() == options.size()) {
+    return {
+        make(sourcemeta::blaze::InstructionIndex::AssertionEqualsAnyStringHash,
+             context, schema_context, dynamic_context,
+             to_string_hashes(perfect_string_hashes))};
+  }
+
+  return {make(sourcemeta::blaze::InstructionIndex::AssertionEqualsAny, context,
+               schema_context, dynamic_context, std::move(options))};
+}
+
+auto compiler_draft3_validation_uniqueitems(
+    const Context &context, const SchemaContext &schema_context,
+    const DynamicContext &dynamic_context, const Instructions &)
+    -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_boolean() ||
+      !schema_context.schema.at(dynamic_context.keyword).to_boolean()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  return {make(sourcemeta::blaze::InstructionIndex::AssertionUnique, context,
+               schema_context, dynamic_context, ValueNone{})};
+}
+
+auto compiler_draft3_validation_maxlength(const Context &context,
+                                          const SchemaContext &schema_context,
+                                          const DynamicContext &dynamic_context,
+                                          const Instructions &)
+    -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_integral()) {
+    return {};
+  }
+
+  assert(schema_context.schema.at(dynamic_context.keyword).is_positive());
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "string") {
+    return {};
+  }
+
+  // We'll handle it at the type level as an optimization
+  if (context.mode == Mode::FastValidation &&
+      schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() == "string") {
+    return {};
+  }
+
+  return {make(
+      sourcemeta::blaze::InstructionIndex::AssertionStringSizeLess, context,
+      schema_context, dynamic_context,
+      ValueUnsignedInteger{
+          static_cast<unsigned long>(
+              schema_context.schema.at(dynamic_context.keyword).as_integer()) +
+          1})};
+}
+
+auto compiler_draft3_validation_minlength(const Context &context,
+                                          const SchemaContext &schema_context,
+                                          const DynamicContext &dynamic_context,
+                                          const Instructions &)
+    -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_integral()) {
+    return {};
+  }
+
+  assert(schema_context.schema.at(dynamic_context.keyword).is_positive());
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "string") {
+    return {};
+  }
+
+  // We'll handle it at the type level as an optimization
+  if (context.mode == Mode::FastValidation &&
+      schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() == "string") {
+    return {};
+  }
+
+  const auto value{static_cast<unsigned long>(
+      schema_context.schema.at(dynamic_context.keyword).as_integer())};
+  if (value <= 0) {
+    return {};
+  }
+
+  return {make(sourcemeta::blaze::InstructionIndex::AssertionStringSizeGreater,
+               context, schema_context, dynamic_context,
+               ValueUnsignedInteger{value - 1})};
+}
+
+auto compiler_draft3_validation_maxitems(const Context &context,
+                                         const SchemaContext &schema_context,
+                                         const DynamicContext &dynamic_context,
+                                         const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_integral()) {
+    return {};
+  }
+
+  assert(schema_context.schema.at(dynamic_context.keyword).is_positive());
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  // We'll handle it at the type level as an optimization
+  if (context.mode == Mode::FastValidation &&
+      schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() == "array") {
+    return {};
+  }
+
+  return {make(
+      sourcemeta::blaze::InstructionIndex::AssertionArraySizeLess, context,
+      schema_context, dynamic_context,
+      ValueUnsignedInteger{
+          static_cast<unsigned long>(
+              schema_context.schema.at(dynamic_context.keyword).as_integer()) +
+          1})};
+}
+
+auto compiler_draft3_validation_minitems(const Context &context,
+                                         const SchemaContext &schema_context,
+                                         const DynamicContext &dynamic_context,
+                                         const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_integral()) {
+    return {};
+  }
+
+  assert(schema_context.schema.at(dynamic_context.keyword).is_positive());
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "array") {
+    return {};
+  }
+
+  // We'll handle it at the type level as an optimization
+  if (context.mode == Mode::FastValidation &&
+      schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() == "array") {
+    return {};
+  }
+
+  const auto value{
+      schema_context.schema.at(dynamic_context.keyword).as_integer()};
+  if (value <= 0) {
+    return {};
+  }
+
+  return {make(sourcemeta::blaze::InstructionIndex::AssertionArraySizeGreater,
+               context, schema_context, dynamic_context,
+               ValueUnsignedInteger{static_cast<unsigned long>(value - 1)})};
+}
+
+auto compiler_draft3_validation_maximum(const Context &context,
+                                        const SchemaContext &schema_context,
+                                        const DynamicContext &dynamic_context,
+                                        const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_number()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "integer" &&
+      schema_context.schema.at("type").to_string() != "number") {
+    return {};
+  }
+
+  // TODO: As an optimization, if `minimum` is set to the same number, do
+  // a single equality assertion
+
+  assert(schema_context.schema.is_object());
+  if (schema_context.schema.defines("exclusiveMaximum") &&
+      schema_context.schema.at("exclusiveMaximum").is_boolean() &&
+      schema_context.schema.at("exclusiveMaximum").to_boolean()) {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionLess, context,
+                 schema_context, dynamic_context,
+                 sourcemeta::core::JSON{
+                     schema_context.schema.at(dynamic_context.keyword)})};
+  } else {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionLessEqual,
+                 context, schema_context, dynamic_context,
+                 sourcemeta::core::JSON{
+                     schema_context.schema.at(dynamic_context.keyword)})};
+  }
+}
+
+auto compiler_draft3_validation_minimum(const Context &context,
+                                        const SchemaContext &schema_context,
+                                        const DynamicContext &dynamic_context,
+                                        const Instructions &) -> Instructions {
+  if (!schema_context.schema.at(dynamic_context.keyword).is_number()) {
+    return {};
+  }
+
+  if (schema_context.schema.defines("type") &&
+      schema_context.schema.at("type").is_string() &&
+      schema_context.schema.at("type").to_string() != "integer" &&
+      schema_context.schema.at("type").to_string() != "number") {
+    return {};
+  }
+
+  // TODO: As an optimization, if `maximum` is set to the same number, do
+  // a single equality assertion
+
+  assert(schema_context.schema.is_object());
+  if (schema_context.schema.defines("exclusiveMinimum") &&
+      schema_context.schema.at("exclusiveMinimum").is_boolean() &&
+      schema_context.schema.at("exclusiveMinimum").to_boolean()) {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionGreater, context,
+                 schema_context, dynamic_context,
+                 sourcemeta::core::JSON{
+                     schema_context.schema.at(dynamic_context.keyword)})};
+  } else {
+    return {make(sourcemeta::blaze::InstructionIndex::AssertionGreaterEqual,
+                 context, schema_context, dynamic_context,
+                 sourcemeta::core::JSON{
+                     schema_context.schema.at(dynamic_context.keyword)})};
+  }
+}
+
+} // namespace internal
+#endif
